Remove the pod configuration for prometheus
I decided these are not services that are bound to each other the way a
database is to a service. I'm going to keep them separate, since it
doesn't really change much. This way, if I ever end up with two
exporters that want to use the same port, I won't have an issue.